リザルト用のデータ当てはめ, 画像の追加, ResultManagerプレハブ作成, ランク機能実装

+++ TGSProject/Assets/Scripts/niitsuma/Results/ResultData.cs
@@ -1,7 +1,7 @@
 ﻿
 public class ResultData
 {
-    public float playTime = 0;
+    public int playTime = 0;
     public int ownHerb = 0;
     public int killEnemyCount = 0;
     public int stoneMonumentCount = 0;
